Refactored app.py, created new file for LogCrew
